refactored the Kwf_Tabs component, moved its .js and .scss files to a new location within the kfw component folder
